검토 명령 서비스를 단일 서비스로 단순화
